Printf warning elimination, fixed a segfault in task_view

+++ src/common.h
@@ -14,7 +14,7 @@ typedef unsigned int u32;
 typedef unsigned short u16;
 typedef unsigned char u8;
 typedef long long s64;
-typedef long s32;
+typedef int s32;
 typedef short s16;
 
 // TODO strong typedefs?
